Update to GENOA v2.0

+++ src/files/ssh-aerosol-genoa/src/include/SOAP/aiomfac.cxx
@@ -20,14 +20,12 @@ void aiomfac_ssh(Array<double, 1> &X_solvents, Array<double, 1> &gamma_LR_solven
   int nsolvents=X_solvents.size();
   int nions=molality_ions.size();
  
-  //long double rho=config.rho_aqueous;
-  long double rho=997.0; //1000.0;  //volumic mass in kg.m-3 
-  long double permittivity=78.54;  //relative static permittivity of water"
+  //double rho=config.rho_aqueous;
+  double rho=997.0; //1000.0;  //volumic mass in kg.m-3 
+  double permittivity=78.54;  //relative static permittivity of water"
 
   int i,j,k,l;    
-  long double logmax=12.;
-
-  //cout << ionic << " " << Temperature << endl;
+  double logmax=12.;
    
   /*cout << "A" << groups_solvents << endl;
     cout << "B" << b1ki << endl;
@@ -53,13 +51,13 @@ void aiomfac_ssh(Array<double, 1> &X_solvents, Array<double, 1> &gamma_LR_solven
       // LONG - RANGE computation           //
       ////////////////////////////////////////
 
-      long double sqrt_ionic=pow(ionic,0.5);
-      long double sqrt_rho=pow(rho,0.5);
-      long double A=1.327757e5*sqrt_rho/(pow(permittivity*Temperature,3.0/2));
-      long double b=6.369696*sqrt_rho/(pow(permittivity*Temperature,0.5));      
-      long double A2=A*(1+b*sqrt_ionic-1.0/(1+b*sqrt_ionic)
+      double sqrt_ionic=pow(ionic,0.5);
+      double sqrt_rho=pow(rho,0.5);
+      double A=1.327757e5*sqrt_rho/(pow(permittivity*Temperature,3.0/2));
+      double b=6.369696*sqrt_rho/(pow(permittivity*Temperature,0.5));      
+      double A2=A*(1+b*sqrt_ionic-1.0/(1+b*sqrt_ionic)
                    -2.0*log(1.0+b*sqrt_ionic))/pow(b,3);
-      long double A3=A*sqrt_ionic/(1+b*sqrt_ionic);
+      double A3=A*sqrt_ionic/(1+b*sqrt_ionic);
 
       static Array<double, 2> Bki,IBki_deriv;
       static Array<double, 2> Bca,IBca_deriv;
@@ -123,8 +121,8 @@ void aiomfac_ssh(Array<double, 1> &X_solvents, Array<double, 1> &gamma_LR_solven
       
       Bki=0.;
       Bca=0.;      
-      long double I1=exp(-1.2*sqrt_ionic);
-      long double I2=I1/sqrt_ionic;
+      double I1=exp(-1.2*sqrt_ionic);
+      double I2=I1/sqrt_ionic;
 
       for (k=0;k<ngroups;++k)
         for (i=0;i<nions;++i)
@@ -141,7 +139,7 @@ void aiomfac_ssh(Array<double, 1> &X_solvents, Array<double, 1> &gamma_LR_solven
         for (j=0;j<nions;++j)
           if(charges_ions(i)*charges_ions(j)<0.0)
             {
-              long double expb3ca=exp(-b3ca(i,j)*sqrt_ionic);
+              double expb3ca=exp(-b3ca(i,j)*sqrt_ionic);
               Bca(i,j)=b1ca(i,j)+b2ca(i,j)*expb3ca;
               IBca_deriv(i,j)=-0.5*b3ca(i,j)*b2ca(i,j)*
                 expb3ca/sqrt_ionic;
@@ -156,7 +154,7 @@ void aiomfac_ssh(Array<double, 1> &X_solvents, Array<double, 1> &gamma_LR_solven
 	      ICca_deriv(i,j)=0.0;
             }
 	  
-      long double sumXk=0.0;
+      double sumXk=0.0;
       for (k=0;k<ngroups;++k)
         {
           Xk(k)=0.0;
@@ -173,11 +171,11 @@ void aiomfac_ssh(Array<double, 1> &X_solvents, Array<double, 1> &gamma_LR_solven
         cout << "sumXk is zero" << endl;
 
       //cout << Xk << endl;
-      long double molar_mass_average=0.0;
+      double molar_mass_average=0.0;
       for (k=0;k<ngroups;++k)
         molar_mass_average+=Xk(k)*molar_mass_groups(k); //CHANGED Based on a corrigeum published in ACP by Zuend. BEFORE: molar_mass_average+=X_solvents(k)*molar_mass_solvents(k); 
 
-      long double mi_zi=0.0;
+      double mi_zi=0.0;
       for (i=0;i<nions;++i)
         mi_zi+=molality_ions(i)*abs(charges_ions(i));
       
@@ -229,7 +227,7 @@ void aiomfac_ssh(Array<double, 1> &X_solvents, Array<double, 1> &gamma_LR_solven
               
               log_gammak(k)+=Bki(k,i)*molality_ions(i);	                 
               
-              long double a=0.;                 
+              double a=0.;                 
               for (l=0;l<ngroups;++l)            
                 a+=(Bki(l,i)+IBki_deriv(l,i)*ionic)*Xk(l)/molar_mass_average;
               
@@ -263,12 +261,12 @@ void aiomfac_ssh(Array<double, 1> &X_solvents, Array<double, 1> &gamma_LR_solven
       sumIBkXk=0.;
       for (j=0;j<nions;++j)
         {
-          //long double a=0.;
+          //double a=0.;
           for (k=0;k<ngroups;++k)              
             sumIBkXk(j)+=IBki_deriv(k,j)*Xk(k);
         }
 
-      long double loggamma;
+      double loggamma;
       for (i=0;i<nions;++i)
         if(charges_ions(i)>0.0 and (iH<0 or i==iH))
           {
@@ -278,7 +276,7 @@ void aiomfac_ssh(Array<double, 1> &X_solvents, Array<double, 1> &gamma_LR_solven
 			
             for (j=0;j<nions;++j)
               {
-                long double a=sumIBkXk(j);
+                double a=sumIBkXk(j);
 
                 a*=pow(charges_ions(i),2)/(2*molar_mass_average);
 			          
@@ -315,7 +313,7 @@ void aiomfac_ssh(Array<double, 1> &X_solvents, Array<double, 1> &gamma_LR_solven
             //  cout << "A: " << loggamma << " " << molar_mass_average << endl;            
             for (j=0;j<nions;++j)
               {
-                long double a=sumIBkXk(j);
+                double a=sumIBkXk(j);
                 a*=pow(charges_ions(i),2)/(2*molar_mass_average);
                 /*
                 if (i==6)
